chore: finish modularization, separate watcher code
