+++ src/dbt_core_interface/project.py
@@ -46,7 +46,7 @@ from dbt.contracts.graph.manifest import Manifest
 from dbt.contracts.graph.nodes import ManifestNode, SourceDefinition
 from dbt.flags import set_from_args
 from dbt.parser.manifest import ManifestLoader, process_node
-from dbt.parser.read_files import FileDiff, InputFile, ReadFilesFromFileSystem
+from dbt.parser.read_files import FileDiff, InputFile
 from dbt.parser.sql import SqlBlockParser, SqlMacroParser
 from dbt.task.sql import SqlCompileRunner
 from dbt.tracking import disable_tracking
@@ -59,6 +59,7 @@ if t.TYPE_CHECKING:
     from dbt.cli.main import dbtRunnerResult
     from sqlfluff.core.config import FluffConfig
     from sqlfluff.core.linter.linted_dir import LintingRecord
+
 
 disable_tracking()
 set_invocation_context(get_env())
@@ -621,10 +622,6 @@ class DbtProject:
                 return node
         return None
 
-    def create_project_watcher(self, check_interval: float = 2.0) -> DbtProjectWatcher:
-        """Create a project watcher for automatic incremental updates."""
-        return DbtProjectWatcher(self, check_interval)
-
     def command(self, cmd: str, *args: t.Any, **kwargs: t.Any) -> dbtRunnerResult:
         """Run a dbt command with the current project manifest."""
         from dbt.cli.main import dbtRunner
@@ -808,116 +805,3 @@ class DbtProject:
             f"format_command returning success={success}, result_sql={result_sql[:100] if result_sql is not None else 'n/a'}"
         )
         return success, result_sql
-
-
-@t.final
-class DbtProjectWatcher:
-    """Watch dbt files for changes and automatically update the manifest."""
-
-    def __init__(self, project: DbtProject, check_interval: float = 2.0) -> None:
-        self._project = project
-        self.check_interval = check_interval
-
-        self.reader = ReadFilesFromFileSystem(
-            all_projects=self._project.runtime_config.load_dependencies(),
-            files={},
-            saved_files=self._project.manifest.files,
-        )
-
-        self._mtimes: dict[Path, float] = {}
-        self._running = False
-        self._thread: threading.Thread | None = None
-        self._stop_event = threading.Event()
-
-    def start(self) -> None:
-        """Start monitoring files for changes."""
-        if self._running:
-            return
-
-        self._running = True
-        self._stop_event.clear()
-        self._thread = threading.Thread(target=self._monitor_loop, daemon=True)
-        self._thread.start()
-        logger.info("Project watcher started")
-
-    def stop(self) -> None:
-        """Stop monitoring files."""
-        if not self._running:
-            return
-
-        self._running = False
-        self._stop_event.set()
-        if self._thread:
-            self._thread.join(timeout=5.0)
-        logger.info("Project watcher stopped")
-
-    def _monitor_loop(self) -> None:
-        """Run the main monitoring loop."""
-        self._initialize_file_mtimes()
-        set_invocation_context(get_env())
-
-        while self._running and not self._stop_event.is_set():
-            try:
-                change_level = self._check_for_changes()
-                if change_level:
-                    self._project.parse_project(
-                        write_manifest=False, reparse_configuration=change_level > 1
-                    )
-            except Exception as e:
-                logger.error(f"Error in project watcher loop: {e}")
-
-            _ = self._stop_event.wait(self.check_interval)
-
-    def _initialize_file_mtimes(self) -> None:
-        """Initialize the file modification time tracking."""
-        for f_proxy in self._project.manifest.files.values():
-            path = Path(f_proxy.path.absolute_path)
-            if path.exists():
-                self._mtimes[path] = path.stat().st_mtime
-        self._mtimes[self._project.dbt_project_yml] = self._project.dbt_project_yml.stat().st_mtime
-        self._mtimes[self._project.profiles_yml] = self._project.profiles_yml.stat().st_mtime
-        logger.debug(f"Initialized tracking for {len(self._mtimes)} files")
-
-    def _check_for_changes(self) -> int:
-        """Check for changes in tracked files.
-
-        A return value of 0 means no changes, 1 means files were added/removed, and 2 means
-        a configuration file was modified (dbt_project.yml or profiles.yml).
-        """
-        for path in (self._project.dbt_project_yml, self._project.profiles_yml):
-            try:
-                current_mtime = path.stat().st_mtime if path.exists() else 0.0
-                stamped_mtime = self._mtimes.get(path)
-                if stamped_mtime is None or current_mtime > stamped_mtime:
-                    self._mtimes[path] = current_mtime
-                    return 2
-            except OSError as e:
-                logger.warning(f"Error checking file {path}: {e}")
-                continue
-
-        self.reader.read_files()
-
-        changes_detected = 0
-        for k, f_proxy in list(self.reader.files.items()):
-            path = Path(f_proxy.path.absolute_path)
-            try:
-                if not path.exists():  # DELETED
-                    _ = self._mtimes.pop(path, None)
-                    _ = self.reader.files.pop(k, None)
-                    changes_detected = 1
-
-                current_mtime = path.stat().st_mtime if path.exists() else 0.0
-                stamped_mtime = self._mtimes.get(path)
-
-                if stamped_mtime is None:  # ADDED
-                    changes_detected = 1
-                elif current_mtime > stamped_mtime:  # CHANGED
-                    changes_detected = 1
-
-                self._mtimes[path] = current_mtime
-
-            except OSError as e:
-                logger.warning(f"Error checking file {path}: {e}")
-                continue
-
-        return changes_detected
